symphonia_format_mkv

Module sub_fields

Key name constants for sub-fields of MKV tags and chapters.

For the exact meaning of these fields, and the format of their values, please consult the official Matroska specification.

Constants

CHAPTER_TITLE_COUNTRY

CHAPTER_TITLE_LANGUAGE

CHAPTER_TITLE_LANGUAGE_BCP47

EDITION_TITLE_LANGUAGE_BCP47

TAG_LANGUAGE

TAG_LANGUAGE_BCP47